+++ src/atoms/button/component.tsx
@@ -1,15 +1,56 @@
 import React from 'react';
 
-import { getClassName } from '@kibalabs/core';
+import { getClassName, RecursivePartial } from '@kibalabs/core';
 import { Link as CoreLink, OptionalProppedElement, useIsCoreRoutingEnabled } from '@kibalabs/core-react';
 import styled from 'styled-components';
 
 import { IButtonTheme } from './theme';
-import { defaultComponentProps, IComponentProps, LoadingSpinner, themeToCss, useBuiltTheme } from '../..';
+import { defaultComponentProps, IComponentProps, LoadingSpinner, themeToCss } from '../..';
 import { Alignment, getFlexContentAlignment } from '../../model';
 import { IIconProps, PaddingSize, Spacing } from '../../particles';
 import { CssConverter } from '../../util';
 
+export const ButtonThemedStyle = (theme: RecursivePartial<IButtonTheme>): string => `
+  & > .focus-fixer {
+    ${themeToCss(theme.normal?.default?.text)};
+    ${themeToCss(theme.normal?.default?.background)};
+  }
+  /* Since it can be rendered as an <a>, unset everything for visited */
+  &:visited > .focus-fixer {
+    ${themeToCss(theme.normal?.default?.text)};
+    ${themeToCss(theme.normal?.default?.background)};
+  }
+  &:hover > .focus-fixer {
+    ${themeToCss(theme.normal?.hover?.text)};
+    ${themeToCss(theme.normal?.hover?.background)};
+  }
+  &:active > .focus-fixer {
+    ${themeToCss(theme.normal?.press?.text)};
+    ${themeToCss(theme.normal?.press?.background)};
+  }
+  &:focus > .focus-fixer {
+    ${themeToCss(theme.normal?.focus?.text)};
+    ${themeToCss(theme.normal?.focus?.background)};
+  }
+  &.disabled {
+    & > .focus-fixer {
+      ${themeToCss(theme.disabled?.default?.text)};
+      ${themeToCss(theme.disabled?.default?.background)};
+    }
+    &:hover > .focus-fixer {
+      ${themeToCss(theme.disabled?.hover?.text)};
+      ${themeToCss(theme.disabled?.hover?.background)};
+    }
+    &:active > .focus-fixer {
+      ${themeToCss(theme.disabled?.press?.text)};
+      ${themeToCss(theme.disabled?.press?.background)};
+    }
+    &:focus > .focus-fixer {
+      ${themeToCss(theme.disabled?.focus?.text)};
+      ${themeToCss(theme.disabled?.focus?.background)};
+    }
+  }
+`;
 
 // NOTE(krishan711): focus problem fixed with https://www.kizu.ru/keyboard-only-focus/#proper-solution
 
@@ -22,7 +63,7 @@ const StyledButtonText = styled.span<IStyledButtonTextProps>`
 `;
 
 interface IStyledButtonProps {
-  $theme: IButtonTheme;
+  $theme?: RecursivePartial<IButtonTheme>;
   $isLoading: boolean;
 }
 
@@ -50,52 +91,17 @@ const StyledButtonInner = styled.span<IStyledButtonInnerProps>`
 
 
 const StyledButton = styled.button<IStyledButtonProps>`
-  cursor: ${(props: IStyledButtonProps): string => (props.$isLoading ? 'default' : 'pointer')};
   transition-duration: 0.3s;
-
   &.fullWidth {
     width: 100%;
   }
-
-  & > .focus-fixer {
-    ${(props: IStyledButtonProps): string => themeToCss(props.$theme.normal.default.text)};
-    ${(props: IStyledButtonProps): string => themeToCss(props.$theme.normal.default.background)};
-  }
-  /* Since it can be rendered as an <a>, unset everything for visited */
-  &:visited > .focus-fixer {
-    ${(props: IStyledButtonProps): string => themeToCss(props.$theme.normal.default.text)};
-    ${(props: IStyledButtonProps): string => themeToCss(props.$theme.normal.default.background)};
-  }
-  &:hover > .focus-fixer {
-    ${(props: IStyledButtonProps): string => themeToCss(props.$theme.normal.hover?.text)};
-    ${(props: IStyledButtonProps): string => themeToCss(props.$theme.normal.hover?.background)};
-  }
-  &:active > .focus-fixer {
-    ${(props: IStyledButtonProps): string => themeToCss(props.$theme.normal.press?.text)};
-    ${(props: IStyledButtonProps): string => themeToCss(props.$theme.normal.press?.background)};
-  }
-  &:focus > .focus-fixer {
-    ${(props: IStyledButtonProps): string => themeToCss(props.$theme.normal.focus?.text)};
-    ${(props: IStyledButtonProps): string => themeToCss(props.$theme.normal.focus?.background)};
-  }
   &.disabled {
     cursor: not-allowed;
-    & > .focus-fixer {
-      ${(props: IStyledButtonProps): string => themeToCss(props.$theme.disabled.default?.text)};
-      ${(props: IStyledButtonProps): string => themeToCss(props.$theme.disabled.default?.background)};
-    }
-    &:hover > .focus-fixer {
-      ${(props: IStyledButtonProps): string => themeToCss(props.$theme.disabled.hover?.text)};
-      ${(props: IStyledButtonProps): string => themeToCss(props.$theme.disabled.hover?.background)};
-    }
-    &:active > .focus-fixer {
-      ${(props: IStyledButtonProps): string => themeToCss(props.$theme.disabled.press?.text)};
-      ${(props: IStyledButtonProps): string => themeToCss(props.$theme.disabled.press?.background)};
-    }
-    &:focus > .focus-fixer {
-      ${(props: IStyledButtonProps): string => themeToCss(props.$theme.disabled.focus?.text)};
-      ${(props: IStyledButtonProps): string => themeToCss(props.$theme.disabled.focus?.background)};
-    }
+  }
+  cursor: ${(props: IStyledButtonProps): string => (props.$isLoading ? 'default' : 'pointer')};
+
+  && {
+    ${(props: IStyledButtonProps): string => (props.$theme ? ButtonThemedStyle(props.$theme) : '')};
   }
 `;
 
@@ -135,15 +141,14 @@ export const Button = (props: IButtonProps): React.ReactElement => {
     throw new Error('if the buttonType is set to submit, you should not use onClicked. use the form.onSubmitted instead');
   }
 
-  const theme = useBuiltTheme('buttons', props.variant, props.theme);
   const isTargetWithinApp = props.target && props.target.startsWith('/');
   const targetShouldOpenSameTab = props.targetShouldOpenSameTab || props.target?.startsWith('#') || (props.targetShouldOpenSameTab == null && isTargetWithinApp);
   return (
     // @ts-ignore: as prop doesn't match type required
     <StyledButton
       id={props.id}
-      className={getClassName(Button.displayName, props.className, props.isFullWidth && 'fullWidth', !props.isEnabled && 'disabled')}
-      $theme={theme}
+      className={getClassName(Button.displayName, props.className, props.isFullWidth && 'fullWidth', !props.isEnabled && 'disabled', ...(props.variant?.split('-') || []))}
+      $theme={props.theme}
       $isLoading={props.isLoading || false}
       onClick={onClicked}
       disabled={!props.isEnabled}