+++ src/util/index.ts
@@ -1,3 +1,3 @@
 export * from './componentUtil';
 export * from './responsiveUtil';
-export * from './theme';
+export * from './themeUtil';